+++ CluckADuck/Entities/Ducky.cpp
@@ -118,7 +118,7 @@ void Ducky::update(float dt)
 void Ducky::draw(sf::RenderTarget& rt)
 {
 	// Set up transformations.
-	spr.setPosition(pos.x, pos.y);
+	spr.setPosition(static_cast<float>(pos.x), static_cast<float>(pos.y));
 	spr.setRotation(static_cast<float>(atan2(vel.y, vel.x) / 3.141592653589793238462643383279502884 * 180.));
 	spr.setScale(static_cast<float>(scale), static_cast<float>((vel.x>=0.)?scale:-scale));
